+++ src/styles/theme.js
@@ -147,6 +147,14 @@ export const defaultTheme = {
       reactionCount: {},
       text: {},
     },
+    actionSheet: {
+      titleContainer: {},
+      titleText: {},
+      buttonContainer: {},
+      buttonText: {},
+      cancelButtonContainer: {},
+      cancelButtonText: {},
+    },
   },
 
   messageInput: {
@@ -189,6 +197,12 @@ export const defaultTheme = {
         top: {},
         title: {},
       },
+    },
+    actionSheet: {
+      titleContainer: {},
+      titleText: {},
+      buttonContainer: {},
+      buttonText: {},
     },
   },
 
@@ -268,6 +282,7 @@ export const themed = (OriginalComponent) => {
               path.pop();
               return { css: v };
             }
+            path.pop();
             return v;
           };